barebox: update to 2016.08.0/customers/skov/imx6/20160914-1
This will add squashfs compression support for lzo, lz4, and zlib
and also extends squashfs support in order to be able to use simple boot
targets and bootspec with appendroot feature.

This will also switches to the new bootchooser implementation which will
require further setup in the environment.

Note: this patchstack also contains the NAND DMA read error fix

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>