BSP: barebox/linux: rename dol63x to imx6
Barebox and the linux kernel still used "dol63x" as the name for the
project/platform - replace it with "imx6". To do so barebox is updated to
2016.08.0/customers/skov/imx6/20161001-1 which includes some bugfixes in the
context of bootspec, subpage read support for the VID header, a change of the
default priority for system0 from 20 to 30, preparation of the state for storing
xres, yres and brightness as well as cutting support for board variants 1-18
besides the renaming mentioned above. The linux kernel is updated to
4.4/customers/skov/imx6/20160930-1 which also renames the platform, cuts support
for board variants 1-18 and implements subpage read support for VID headers.

All known recipes, configs etc. have been adjusted to the new naming so that
everything should be consistent. Unfortunately, this consistence required doing
everything at once so that the result is a rather big and a little confusing
commit.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>