machine: squashfs lzo compression options update
Ulrich noticed that we missed a few paramters in the extra
options to mksquasfs (when creating a bundle).
Added the options so a squashfs bundle uses the same parameters
as used when creating a full image.

When creating the full image the extra options are part of the
genimage configuration.