barebox: update to 2016.11.0/customers/skov/imx6/20161123-1
This release includes the fix for a sporadic SPI issue that hang the boards from
time to time. The basic barebox version has been updated to v2016.11.0 which
already ships support for bootchooser and contains several bugfixes so that the
patch stack shrank substantially compared to the previous release. Additionally
Skov's patch to support board variants 21 & 22 has been integrated into
Pengutronix' local barebox repo and will be squashed into the patch for basic
platform support in the next release.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>