barebox: env: splash: show splash screen on HDMI display
In addition to showing a splash screen the brightness of the display has to be
adjusted (this is because the Chinese HDMI converter not only sets a brightness
but also saves it for usage after the next power cycle). The value to set is
taken from state.display.brightness.

The i2c message necessary for changing the brightness is constructed according
to the monitor control command set (MCCS) and needs barebox' i2c_write command
capable of accepting decimally formatted data. This currently requires a patch
which is sent upstream.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>