barebox: update to 2017.03.0/customers/skov/imx6/20171020-1
This release mainly allows usage of the HDMI display from within barebox which
is necessary to show a splash screen there (task 010.660). Other than in the
case of a parallel display that is darkened right before barebox hands over
control to the linux kernel this is not done in the HDMI case. The reason for
that is that the HDMI display seems to be lazy and slow and starts to show the
splash screen only after linux is started. The risk is that the displayed
framebuffer is filled with binary garbage until linux userspace displays its own
splash screen (although this has not been seen at all while developing the
feature).

Showing a splash screen on the HDMI display needs a change in the script
'/env/init/splash' thus a change in the default environment. As it is assumed
that nobody ever changed this file locally in his environment the value of the
environment magic is not touched.

To be able to control the display brightness via specially constructed I²C
messages the 'i2c_write' command has to be allowed to parse decimally formatted
numbers. It is not clear up to now if the related patch will go upstream.

In the context of the controller's Chinese HDMI converter presumably booting up
too slowly a bug in barebox came apparent which is fixed by a patch from Sascha
Hauer.

Last but not least the order of including barebox' devicetree snippets has been
corrected and an alias node for the state has been introduced as it is needed
starting with v2017.06.0.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>