linux-skov: update to 4.16/customers/skov/imx6/20180425-1 (4.9.96 -> 4.16.4)
Besides upgrading to the newest stable kernel this release enables the usage of
etnaviv as the basis for applying QML and displaying videos in Qt.

IMPORTANT: This is a pre-release as the topic-branch to support the micrel-smi
is missing because it needs additional care. Hence do not forget to re-enable
the micrel switch after this has been done (CONFIG_MICREL_SMI=y).
The DT currently configures the system to have a fixed-phy.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>