barebox: boot arm9 kernel
Adapt env/* files to support booting the kernel.
Use "boot sd"

To do this add the detect command.
Dropped all env files from the board in barebox as they
are not used when we include env files in yocto.
And having unused env files in the baord in barebox was just confusing

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>