barebox: fix boot
There seems to be a residual char in the buffer from the serial port
As a workaround for this bring in the init script
and make sure we ignore the first character.

This makes it more difficult to avoid the scripts in init/
but the alternative where we could not boot automagically
was worse.

The init script is a simplified version from
defaultenv-2-base from barebox.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>