linux-skov: update to 4.17/customers/skov/imx6/20180607-1 (4.17-rc5 -> 4.17)
Besides upgrading to the final v4.17 this release also includes the newest
patches concerning etnaviv & coda. Additionally the recipe reflects that
compared to Linus Torvald's release candidates the vanilla kernels are
downloaded from a different URL.

IMPORTANT: do not forget that currently the topic-branch to support the
micrel-smi is missing because it needs additional care. Hence do not forget to
re-enable the micrel switch after this has been done (CONFIG_MICREL_SMI=y). The
DT currently configures the system to have a fixed-phy.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>