conf: drop devicetree for arm9-cpu
For arm9-cpu the devicetree is passed from barebox to the
kernel, so no need for a dedicated device tree.
Remove the device tree from the configuration, and drop
the unused bootspec from the images as this implicitly
pulled in support for a kernel device tree, and it was
anyway unused for arm9-cpu.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>