Write i.MX6's OCOTP from Linux Userspace (#27)
* barebox: update to 2018.04.0/customers/skov/imx6/20180717-1 (v2018.04.0 -> v2018.04.0)

Besides including sra@SKO's arm9-topic-branch into Pengutronix'internal patch
management this release repairs showing the splash screen on the 10-inch HDMI
display: the Chinese HDMI converter board (used to connect the 10-inch-display)
is known to boot up too slowly, hence it needs longer than specified to get
ready for answering barebox' I2C query for the EDID. To not come back with no
valid modes (resulting in a failure to show the splash screen) implicitely allow
barebox to wait longer by considerably increasing the number of retries. Let
this increase depend on the "display.external" state variable's content to not
catch a noticeable penalty in boot time for a controller with parallel panel.

The provisory patch of Phytec's Christian Hemp to fix a NAND issue ("mtd:
nand_mxs: fix NAND error when change clk rate") was replaced by a patch of
Sascha Hauer ("mtd: gpmi-nand: Make sure clock is disabled during rate change")
that enhanced the former and is going to be mainlined.

Starting with commit 1201055ad53e ("scripts/dtc: Update to upstream version
1.4.6") included in barebox v2018.07.0 the DT-compiler gave new warnings in the
context of the devicetrees customized for the i.MX6 controllers so that they
have been adjusted accordingly.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>

* linux-skov: update to 4.17/customers/skov/imx6/20180713-1 (4.17 -> 4.17.6)

Next to updating to the currently most recent stable release v4.17.6 and
including the newest coda patches the main focus of this release is adding the
property "read-only" to the OCOTP's devicetree node to make it impossible to
burn fuses by default. To still program the OCOTP one needs to explicitely
remove this property which can be done for example from within barebox using a
DT-fixup constructed via

  barebox@Skov IMX6:/ of_property -d -f /soc/aips-bus@2100000/ocotp@21bc000 read-only

While already modifying the DT the display node has been renamed from
"display@di0" to "display" to remove a warning of the DT compiler that has been
seen first with version 1.4.6.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>

* barebox-sd: re-enable burning the OCOTP's fuses via a DT fixup

This is necessary to burn the controller's future MAC address from Linux
userspace.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>

* README: add instructions to fuse the OCOTP from Linux userspace

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>