barebox: fix patches to be git am compatible
Some of the patches failed when using git am.
This breaks the workflow when the patches are added to an external git
tree and then further changes are made there.
It may also break the workflow used by Pengutronix when updating to a
later barebox version.
Replace the patches with variants that are git am compatible.

No functional changes.