linux-skov: update to 5.2/customers/skov/imx6/20190815-1 (5.0.5 -> 5.2.8)
Updating to a v5.2 based kernel was more elaborative than anticipated as Andrew
Lunn's branch 'dsa-tag-modules' which was merged in v5.2-rc1 (see [1]) changed
things considerably and therefore generated many conflicts. In addition updating
the kernel's configuration took more time than expected as the update touched
many options that are used in this BSP.

On the other hand the former topic-branch 'customers/skov/imx6_backports_from_
upstream' is not needed anymore as its patches are found in v5.2 already.

[1] 5f0d736e7f7d..c1a92c0a6c56