linux-skov: update to 5.4/customers/skov/imx6/20200109-1 (5.2.8 -> 5.4.8)
Updating to a v5.4 based kernel brought an updated topic branch 'topic/ksz8873'
which is on its way to being mainlined. Our graphics team suggested to add the
fixes from the 'topic/drm-sched' branch and on the other hand the former
topic-branch 'topic/touchscreen-ads7846' is not needed anymore as its contents
went upstream and can be found in v5.4 already.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>