linux-skov: update to 5.4/customers/skov/imx6/20200331-1 (5.2.21 -> 5.4.28)
Updating to a v5.4 based kernel brought an updated topic branch 'topic/ksz8873'
which is on its way to being mainlined. Our graphics team suggested to add the
fixes from the 'topic/drm-sched' branch and on the other hand the former
topic-branch 'topic/touchscreen-ads7846' is not needed anymore as its contents
went upstream and can be found in v5.4 already.

On the stable branch, shortly after v5.4.5 commit [1] was introduced into the
stable v5.4.y branch that unfortunately breaks the handling of the used SD card
from ATP:

    root@house1-dol53x63xsparepart-01-a0-2e:~# vcmd2702 /dev/mmcblk0
    sdhci-esdhc-imx 2198000.usdhc: Card stuck in wrong state! card_busy_detect
    status: 0xb00
    Error 110
    Failed to send command: Connection timed out

So it was decided by san@SKO to revert it for the moment and report the problem
to ATP. As this is only a temporary workaround that problem really has to be
solved to be future-proof.

With respect to the kernel's build time configuration all modifications from [2]
have been applied to this kernel as well.

[1] 89c6e8829469 ("mmc: block: Add CMD13 polling for MMC IOCTLS with R1B response")
[2] 2f17c6265e20 ("[CORE-5929] imx6-cpu/defconfig: Set unused filesystem as modules             add support for f2fs")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>