linux-skov: update to 5.4/customers/skov/imx6/20200429-1 (5.4.32 -> 5.4.35)
Support for board revision 'e' does not have the work-in-progress status anymore
which is reflected in this customized kernel's name compared to the name of its
predecessor. This is connected with a slight modifikation in the compatible
string for board revision 'e' and the name of its devicetree by introducing the
panels label requiring the updated barebox from the preceding commit.

Next to that an updated topic branch for the micrel switch is included as well
which is a backport of the current version developed on top of v5.7-rc2
(hopefully going to be mainlined soon). It contains a bugfix addressing shortcut
detection which was broken on v5.4 up to now.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>