barebox: update to 2020.06.0/customers/skov/imx6/20200909-1 (2020.06.0 -> 2020.06.0)
To be able to use the Micrel switch's external ports independently (without
being combined into a hardware assisted software bridge) each of them needs a
unique MAC-address. As in that setup the fec (eth0) is unable to send packets to
the outside world one can re-use its MAC-address for the first port (eth1). Look
up the MAC-Address for the second port (eth2) in barebox' state. If it is not
changed from its default value in the devicetree (00:11:22:33:44:55) then
transplant the fec's MAC-address to the second port as well (either both ports
are combined into a bridge or they are connected to different subnets so that a
doubled MAC-address does no harm as well). Finally provide them to Linux by
fixing up the kernel devicetree.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>