linux-skov: update to 5.4/customers/skov/imx6/20200909-1 (5.4.51 -> 5.4.63)
This release persists the provisorial commits [1], [2] & [3] by importing them
into Pengutronix' internal patch management system. Within the devicetree the
reset GPIO is moved from the fec to the switch (see email [4]). Commits [5] &
[6] which entered the LTS kernel branch at v5.4.52 & v5.4.53 let the micrel
switch fail with respect to the current state of its driver. Hence they have
been reverted for the moment. This revert will be substituted by a "real"
solution as soon as there is a new release of the driver developed by Michael
Grzeschik.

[1] 63f83695de64 ("[CORE-7063] Fix the the devicetree, so refresh is 60Hz not 30Hz")
[2] f4233e4cf889 ("{CORE-7022] 0106-drm-panel-add-Multi-Innotechnology-MI1010AIT-1CP1.patch: change devicetree acourding to multi-innos suggestions")
[3] ef46d2471439 ("[CORE-7466] Changed description name to mtach SAP names")
[4] From: Michael Grzeschik <mgr@pengutronix.de>
    Subject: Re: [Customers.Skov] Switch-Reset Hardware Changes
    Date: Thu, 3 Sep 2020 09:28:17 +0200
    Message-Id: 20200903072817.GD9726@pengutronix.de
[5] d9b8206e5323 ("net: dsa: microchip: set the correct number of ports")
[6] d55dad8b1d89 ("net: dsa: microchip: set the correct number of ports")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>