linux-skov: enable SOFTLOCKUP_DETECTOR
The kernel lockup detector should be active, as it's low overhead and
potentially provides useful debug output on bugs that cause the CPU to loop in
kernel mode without giving other tasks a chance to run (softlockup), or cause
the CPU to loop in kernel mode without letting other interrupts have a chance to
run (hardlockup).

Both the hardlockup and the softlockup detector should be enabled when
available.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>