linux-skov: disable VT
Virtual terminal support is not needed on most embedded systems, as any possible
displays are used for the main application. On such systems, VTs can lead to
unwanted kernel messages on boot or disturb the image rendered by the
application If the display should show kernel messages (on boot) and/or provide
a text console with local login, this must not be disabled.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>