linux-skov: disable WATCHDOG_HANDLE_BOOT_ENABLED
If the bootloader has enabled the watchdog, the intention is usually to monitor
the whole boot process. So do not let the kernel feed the watchdog but instead
wait for the userspace to take over. Of course, this means that if your
userspace does not start fast enough your device will reboot.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>