linux-skov: update to 5.10/customers/skov/imx6/20210413-1 (5.4.96 -> 5.10.29)
Next to updating to the currently latest release v5.10.29 of the new longterm
support series this customized kernel contains the backport of Michael
Grzeschik's latest version of the patch series to support the Micrel switch. It
is now based upon the kernel's new cable test API and requires an update of
ethtool to at least v5.8 as that offers the new "--cable-test" option.
Additionally it integrates Oleksij Rempel's new topic branch to selftest the
i.MX6's FEC to detect some basic issues like:

  - MAC loopback fail: most probably wrong clock configuration.
  - PHY loopback fail: incorrect RGMII timings, damaged traces, etc.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>