barebox: update to 2021.07.0/customers/skov/imx6/20210720-1 (2020.06.0 -> 2021.07.0)
This barebox release builds upon the upstream support found in v2021.05.0 for
the first time. While upstreaming the ARM9 platform Sascha adjusted the board's
image's name ("barebox-skovarm9cpu.img" vs. "barebox-skov-arm9cpu.img") which
needs to be reflected in the machine's and genimage's configurations, as well.

For the IMX6 platform the last customized non-upstream release still contained
two files for barebox' environment that did not go mainline:

  env/init/automount
  env/nv/boot.default

While the second one is overwritten by the environment found in the Yocto-BSP
this is not true for the first one. Hence, it is imported from the last
customized non-upstream barebox release here:

  $ ORIGIN=arch/arm/boards/skov-imx6/env/init
  $ DESTINATION=recipes-bsp/barebox/files/imx6-cpu/env/init
  $ mkdir -p ${DESTINATION}
  $ git -C <PATH_TO_BAREBOX_REPO> archive 2020.06.0/customers/skov/imx6/20200909-1:${ORIGIN} automount | \
        tar -C ${DESTINATION} -x

Despite the compactness and convenience of a real defconfig switch to the
detailed buildtime configuration for the ARM9, too, to not implicitely rely upon
the dependencies defined in all the Kconfig files which potentially change from
release to release.

Compared to the previous customized non-upstream barebox releases on the IMX6
platform the board's compatible is now already adjusted in barebox' devicetree.

Another new feature is that the kernel's devicetree is fixed up with the
properties

  /chosen/skov,imx6-board-version
  /chosen/skov,imx6-board-variant

containing the corresponding information.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>