linux-skov: update to 5.10/customers/skov/imx6/20211206-1 (5.10.69 -> 5.10.83)
While updating there were new conflicts within topic branch

  v5.10/topic/ksz8873

that needed to be resolved. Next to that this kernel contains working
multi-bridge support for the Micrel switch (see patch description included in
patch file [1]).

Changes in the kernel's defconfig are the outcome of Linux' commits [2]
contained in v5.10.75 and [3] contained in v5.10.76 for the first time.

[1] 0543-net-dsa-microchip-implement-multi-bridge-support.patch
[2] 6da9af2d2531 ("r8152: select CRC32 and CRYPTO/CRYPTO_HASH/CRYPTO_SHA256")
[3] c56c801391c3 ("ARM: 9122/1: select HAVE_FUTEX_CMPXCHG")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>