barebox: update to 2022.02.0/customers/skov/imx6/20220310-1 (v2021.11.0 -> v2022.02.0)
Next to switching to barebox' latest release this update integrates devicetree
changes that equip the imx8eval platform with a barebox-state that is placed
onto the eMMC.

Additionally the buildtime configuration option MACHINE_ID (see [1]) has been
enabled paired with the introduction of barebox' environment variable

  bootm.provide_machine_id = true

to provide the Linux userspace with a system-individual, unique and particularly
persistent machine-id utilizing the "systemd.machine_id=" kernel command line
parameter.

One problem that is fixed in this release, as well, is that HDMI is now working
again (regarding barebox' splash screen) and that it does not take a 30 s period
before hitting a timeout and finally booting the Linux kernel. The problem was
that barebox performed a workaround for broken PFDs twice: once in the clock
driver and once in architecture code. Deep probe changed the order in which the
workarounds were executed. Together with other bugs this left the HDMI clock
disabled. For more details see [2].

Last not least to ease development the buildtime configuration option SYSCON_
REBOOT_MODE has been enabled which allows to control the bootsource for the
upcoming reboot by manipulation of undocumented general purpose registers that
the i.MX6's boot ROM's code includes into its decision from where to get the
bootloader, see [3]. Independent of the state of jumper J802 one can now enforce
the next boot process to come from the SPI NOR via

  barebox@Skov IMX6:/ global.system.reboot_mode.next=spi0-2
  barebox@Skov IMX6:/ reset

or from the SD-card by

  barebox@Skov IMX6:/ global.system.reboot_mode.next=mmc3
  barebox@Skov IMX6:/ reset

because those registers' contents survive a warm reset.

If the value of variable "system.reboot_mode.next" is persisted using barebox'
respective environment the system will reboot from the desired medium until the
general purpose registers are reset

  barebox@Skov IMX6:/ global.system.reboot_mode.next=normal
  barebox@Skov IMX6:/ reset

or the system is power cycled. Unfortunately there has been a bug in barebox'
reboot mode driver which is now fixed in this release, see [4]: it was
responsible for the failing reset/reboot of some boards leading to a hanging
system, see the thread starting with [5].

[1] https://git.pengutronix.de/cgit/barebox/tree/common/Kconfig?h=v2022.02.0#n1044
[2] http://lists.infradead.org/pipermail/barebox/2022-March/038632.html
[3] https://www.barebox.org/doc/latest/user/reboot-mode.html
[4] http://lists.infradead.org/pipermail/barebox/2022-March/038651.html
[5] From: Søren Andersen <san@skov.dk>
    Subject: [Customers.Skov] Barebox 2022.02
    Date: Thu, 3 Mar 2022 06:38:45 +0000
    Message-Id: AM0PR0402MB34123A8D78641EC0727115B4A6049@AM0PR0402MB3412.eurprd04.prod.outlook.com

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>