rauc: add a bootloader slot for the imx8eval
As barebox needs to be stripped its front-padding before being flashed onto the
eMMC's boot partitions the corresponding feature needs to be backported to
meta-rauc's dunfell branch which is done in pullrequest [1]. Until that is
merged we provisionally add that feature to the BSP by importing the
pullrequest's write_manifest() function.

NOTE: That function must be removed again after [1] is merged and the BSP's
meta-rauc submodule is updated accordingly!

[1] https://github.com/rauc/meta-rauc/pull/257 ("Allow Skipping Bytes from Image")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>