bsp_imx6-cpu.ref, bsp_imx8eval.ref: add bsp-specific rule 'skov_kernel_namespaces'
As the native rule 'kernel_namespaces' only allows to check for all analyzed
namespaces to be off this rule is satisfied by having changed its expected value
to true. A side-effect is that now one does not recognize anymore if one of the
other namespaces is enabled, as well. Therefor add another more specific local
rule that only checks for the remaining namespaces.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>