recipes-support: factory-install: visualize status of installation via LED D2
While D2 is off by default it now shows a nervous flickering (with a period of
50ms) during the factory installation and borrows D1's heartbeat on success or
blinks slowly (with a period of 1s) in the case of an installation failure.

Note: after a successful factory installation D1 and D2 are both shut off before
switching them to heartbeat to have them blink in sync.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>