canned-wks: utilize the empty-plugin for empty partitions
This plugin is available starting with Poky's hardknott release and explicitly
fills the corresponding partitions with a hole leading to a sparse unix file
which is later optimally exploited by the usage of bmaptool.

WIC formats all partitions which do not have a "--source" parameter with the
filesystem specified by the "--fstype" parameter and assigns the identifier
specified by the "--label" parameter. Utilizing the empty plugin those
parameters are useless and thus can be removed now.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>