canned-wks: switch from MBR- to GPT-partitioned SD-card images for the imx8-cpu
Unfortunately this can not easily be done for the imx6-cpu as a GPT is stored in
sectors 0-33 by default which conflicts with the i.MX6's expectation of where
the bootloader is found.

The way out for the i.MX6 is to move the GPT's partition table from sectors 2-33
to the end of the unpartitioned area in front of the first partition (sectors
2016-2047) which could be accomplished using the sgdisk-utility:

  $ sgdisk --adjust-main-table=2016 <SD-CARD-IMAGE.wic>

After this is done barebox could be dd'd to the image:

  $ dd if=<BAREBOX.img> of=<SD-CARD-IMAGE.wic> bs=1024 seek=1 skip=1

In the long run this could be integrated into wic e.g. by addition of a new
parameter "--adjust-main-table=<XY>" to the "bootloader" command.

Utilization of a GPT has the subtle charme that using partition type UUIDs in
combination with a newer systemd allows for automatic mounting of e.g. /home by
systemd-gpt-auto-generator without the need for an entry in /etc/fstab so that
the WKS-file's "--ondisk" and "--fsoptions" parameters could be dropped, too.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>