canned-wks: drop the "--label" options in GPT-partitioned images
Unfortunately for GPT partitioning wic shows the issue that a partition that is
assigned a partition name using the "--part-name" option and that is filled with
a filesystem which shall be assigned a filesystem label via the "--label" option
gets that filesystem label as partition name, as well. This results in no way to
have different partition names/filesystem labels at the moment.

As filesystem labels can still be changed later by the installation of RAUC
bundles let's go without them in the WKS files to allow for defined partition
names in the GPT instead.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>