images: allow GPT partitioning for the imx6-cpu
The i.MX6 expects the bootloader to be stored from sector 2 onwards which
collides with the standard location of the GPT partition table. Hence move the
latter right in front of barebox' environment partition (see [1]) and only
afterwards integrate the barebox binary into the WIC image. Unfortunately WIC
does not support that (at the moment) so do it by postprocessing the WIC image.

[1] https://man.archlinux.org/man/sgdisk.8#j,

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>