[WIP CHG] FIX!! Barebox on LVDS board autoprobe hdmi via barebox env
On sd-card barebox env dont work, which result the hdmi autoprobe will fail if
there are no HDMI, which result in a non bootable board.
Make a quick fix for skov production, which are going to produce 60
boards tomoorrow.
Note SD-boots with no env now result in a LVDS tds, but we can live with
that
Discuss the patch and if we can find a smoother solution.