barebox-sd: add creation of a barebox-environment
It currently only contains the environment variable autoboot_timeout with a
value big enough to allow labgrid to stop barebox from booting. This enables
labgrid to issue commands on the barebox-prompt.

The idea is that this external environment can be injected into an SD-card's
environment partition if required to have barebox' countdown last a little
longer.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>