at91bootstrap: drop from BSP
Now that at91bootstrap has been completely replaced by barebox running
as first stage bootloader, we can rip out the recipe and everything
related to it.

Devices deployed in the field will continue to boot with at91bootstrap,
but for in-field recovery, we would like a SD-Card recovery image that
we build from source, but this is no longer possible with at91bootstrap:

New versions drop support for older ARM9 SoCs like our AT91SAM9263 and
older versions are miscompiled by new toolchains, which causes booting
from SD to hang.