barebox: add SD first stage support for ARM9 CPU
at91bootstrap v3.8.7 fails to boot the second stage at runtime when
built with a new toolchain and newer at91bootstrap versions don't
support the older ARM9 CPU.

To resolve this, Sam had already started to add support for barebox as
first stage bootloader[1] on the AT91SAM9263 reusing infrastructure that
had previously been added for the newer SAMA5D3.

This imports Sam's patch and fixes various smaller and bigger issues to
allow barebox to replace AT91Bootstrap on this platform when booted from
SD-Card. NOR-Flash can be read and written, but is not configured as boot
medium and Linux can be booted.

This will be used for the SD recovery image. Notably missing
functionality is Ethernet and USB in barebox. These don't work with an
up-to-date Linux using the same device tree either, so the root cause
may be the same. Given that this didn't work with AT91Bootstrap either
(when compiled with an older toolchain), this is deemed acceptable.

The patches, except for the final patch prefixed with HACK have all been
submitted to the mailing list already for upstream inclusion.

[1]: https://lore.barebox.org/barebox/20220628203849.2785611-12-sam@ravnborg.org/