barebox: imx8-cpu: decrease VDD_SOC and VDD_ARM to 850 mV
Unlike Linux, the PMIC setup for the imx8-cpu isn't done in the device
tree, but in very early C-Code as this needs to be done while running
out of on-chip SRAM before DRAM setup can even be attempted.

The PMIC configuration we had was for overdrive mode, but we want to run
the SoC in nominal mode instead as that's what the heat sink was specced
for. Nominal mode is 850mV, so configure BUCK1 RUN voltage to exactly
that value.

This alone is however not sufficient to run in nominal mode as several
clocks will exceed the limits validated for nominal drive mode.

Addressing this via device tree configuration like done with Linux
unearthed a bug in the i.MX8M composite clock driver that's fixed here
as well.

We already do some clock configuration while running the prebootloader
in on-chip SRAM. These are tuned a bit for better efficiency.