linux-skov: imx8-cpu: reduce VDD_SOC and VDD_ARM to 850mV
As foretold in commit c500ed3fe035 ("linux-skov: imx8-cpu: dts: fix Linux
setting VDD_3V3 to 3v4"), the voltage for VDD_SOC / VDD_CPU are
warranting a better look. VDD_SOC must be 850mV as that's what the heat
sink was specced for and VDD_ARM should remain subject to dynamic
voltage and frequency scaling. The DVFS range is from 1.2 GHz @ 850mV
to 1.6 GHz @ 950mV, with the lower value being the default.

In general, the i.MX8MP can go up to 1.8GHz at 1.0v, but Linux will mask
out this OPP if the hardware doesn't support it, which is the case for
the imx8-cpu I have here.

The 1.6 GHz operating point needs to be explicitly requested through the
userspace governor, so we don't need to change much in that DT node
except to clamp it to the allowed values for documentation purposes.

VDD_SOC on the other hand is the messier can of worms, as running at
850mV (called nominal mode) imposes restrictions on the maximum clock
frequencies allowed across the SoC. Unfortunately, the upstream support
in Linux (and in barebox) assumes overdrive mode with VDD_SOC of 950mV,
so we need to dial down a number of clocks to remain within the
permissible range, lest strange issues at runtime occur.

The Linux side is done in this patch and a sanity check is added to the
i.MX8MP clock controller driver to avoid runtime clock reconfiguration
that would violate the limits of the nominal drive mode.

With these changes, a single clock remains above nominal frequency: The
GIC clock is running at 500 MHz instead of 400 MHz at Linux kernel
startup and it can't be reconfigured during kernel runtime, because:

  - The clock needs to change parents
  - The mux needs to be gated, because the mux isn't glitch-free
  - The GIC clock is marked critical, so it's not possible to gate it

The workaround for this is to reconfigure the GIC clock in barebox,
where it's possible safely the interrupt controller isn't yet used.
This follows in the next commit.