barebox: imx8-cpu: enable poller command
barebox interleaves a number of operations, called pollers,
periodically with normal execution. We had an issue that should be
fixed now, where a poller that checks network link state took a lot
longer than expected, because the switch IC was misbehaving.
This should be resolved now, but in case similar issues pop up
again, the poller command can aid in debugging.