barebox: imx8-cpu: enable ARM Crypto Extension SHA acceleration
We already enable SHA1 as it's used by machine id generation, but we
use the generic C version, which is much slower than the dedicated
instructions in the ARMv8 Crypto Extension. We don't do much SHA1 at
the moment, but later when doing verified boot, we'll be doing
SHA256 over multiple megabytes of kernel image and having the crypto
extensions will make a noticable difference on boot time.