barebox: imx8-cpu: enable stack guard page
The 32K of barebox stack space is plenty and we haven't observed issues.
If for some reason, the stack is depleted however, the behavior would be
completely erratic, because the stack overflows into the MMU page tables.
Enabling this option, places a page before stack bottom, so an overflow
is caught immediately.