barebox: imx8-cpu: generate fallback Ethernet address from SoC ID
We already use the i.MX SoC ID to derive a machine ID that stays the
same across reboots and updates and is unique for each machine.

barebox can also use the SoC ID to derive fallback MAC addresses in case
no MAC addresses are fused. This is useful for us at Pengutronix,
because otherwise barebox would randomize the address on network boots,
which can lead to depletion of the DHCP pool when a boot loop runs
unchecked.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>