linux-skov: imx8-cpu: enable necessary pstore options in kernel
pstore was already enabled, but not ramoops, which is the pstore backend
we are going to use. Enable it and while at it, add support for pmsg
and console logging. pmsg is for userspace to write messages into
/dev/pmsg0 and console is for storing kernel and barebox console
log. This is independent from dmesg, which is the format used for
reporting panics.