barebox: imx8-cpu: enable pstore
Now that barebox pstore support has been fixed for 64-bit platforms,
enable it. This way barebox log is now readable from userspace in
/sys/fs/pstore and kernel log is appended.

This works fine, but the other way round: seeing kernel log in barebox
and in a subsequently booted kernel is more tricky: Because we do a cold
reset, where DRAM is powered off for 250ms, we get a lot of bitflips.

This will have to be tackled in a separate PR though. Possible solutions
might be moving the pstore area to on-chip SRAM or increasing ECC size.
Turning off the discharge resistors or trying to set bits, so the DRAM
rail is not power cycled was not successful.