barebox: imx8-cpu: import fixes for pstore
barebox pstore support was initially ported for ARM32 and due to a bug
was incompatible with 64-bit Linux. This is fixed upstream in barebox
now, so import the patch as well as support to delete pstore in barebox.

This is useful for testing pstore, but shouldn't be needed in normal
use, because Linux/systemd will take care to persist and clear the
ramoops files.