barebox: imx8-cpu: setup CAAM in barebox
Linux currently fails to initialize the SoC's crypto accelerator (CAAM):

  May 08 14:51:28 imx8-cpu kernel: caam 30900000.crypto: Entropy delay = 3200
  May 08 14:51:28 imx8-cpu kernel: caam 30900000.crypto: failed to acquire DECO 0
  May 08 14:51:28 imx8-cpu kernel: caam 30900000.crypto: failed to instantiate RNG

It's unclear why, but we can work around this by doing the
initialization in barebox, so Linux can doesn't need to fail while
attempting to do so.

It'd still be good to understand why Linux fails as it works on other
i.MX8M SoCs. I'll try to see if this issue always existed on the EVK,
but for SKOV IMX8-CPU purposes, let's just have barebox do the
initialization for now, so we have faster randomness generation.

When/If we do secure boot in future, this option will be enabled anyway,
as it ensures that barebox initialized the CAAM very early, so both
OP-TEE and Linux can make use of the configured job rings.