linux-skov: imx8-cpu: disable CAAM hashing acceleration
Tests showed that ARMv8-A crypto extensions as implemented on the ARM
Cortex-A53 on the i.MX8M family of processors are substantially faster
than the same primitives offered by the CAAM crypto accelerator.

This had a noticeable effect on dm-verity performance when sha256 is used,
e.g. when doing rauc install tc-ci-bundle-imx8-cpu.raucb (verity):

  sha256-caam: 39 seconds
  sha256-ce:   32 seconds

These measurements were done on the IMX8-CPU running in nominal drive
mode with 1.2 GHz CPU frequency.

Besides SHA1/224/256/384/512, the CAAM also accelerates MD5, XCBC(AES)
and CMAC(AES). All of these are also supported by ARMv8 Crypto
Extensions, except for MD5, which we shouldn't use anyway.

Therefore just disable using the CAAM for hashing. Another way would be
to bump down the CAAM priority with respect to the Crypto Extensions,
but this solution saves us having to worry about upstreaming that patch.

Note that this has no effect yet, because the CAAM fails to probe. With
the follow-up commit, the CAAM probe will succeed, which is nice because
of its random number generator, but we don't want to use its slow
hashing accelerator, which is why we do this config change here.