barebox: allow autoboot_timeout adjustment via BAREBOX_AUTOBOOT_TIMEOUT
labgrid is not able to stop barebox from auto booting reliably if the
timeout is too short. In order to allow automating and testing the
images, use the timeout provided in the variable
BAREBOX_AUTOBOOT_TIMEOUT. If the variable is not defined, default to the
value from env/nv/autoboot_timeout.