linux-skov: run reboot notifiers on forced reboot as well
We currently use hw_protection=reboot to reboot instead of shutdown when
mains fail is detected. The HW protection action for shutdown is as
follows:

  - Try orderly shutdown
  - If that times out, try forced shutdown
  - If that fails, try emergency reset

For reboot, the behavior looked like this:

  - Try orderly reboot
  - If that times out, try emergency reset

Notably missing, is trying a forced reboot, so add that back
before falling back to emergency reset (which shouldn't happen).

This was noticed by Jesper because:

  [   18.945660] mmc2: shuting down mmc card
  [   18.949874] mmc: power off notify
  [   18.954352] mmc: power off notify done

was missing. This is because the emergency reset doesn't run the reboot
notifiers, but a forced reset does.

Note that this does _not_ affect watchdog reboots.