trusted-firmware-a: restrict access to secure world memory regions
The main DDR memory can not only be accessed by the CPU but also by various
other bus masters e.g. the SDMA1 engine. To prevent any non-secure R/W access to
the secure world memory regions we have to tag these bus masters as non secure -
which is done via the CSU peripheral in the TF-A. Furthermore we have to protect
the CSU registers that control these access settings to be only changeable by
the secure world.

The patch series that enables this protection has been mainlined into the
upstream TF-A [1] but is not yet included in an official release.

[1] https://review.trustedfirmware.org/q/topic:"feature/imx8m-csu"

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>
Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>